requirements.txt
asgiref==3.9.1
certifi==2025.8.3
Django==5.2.5
django-ckeditor==6.7.3
django-debug-toolbar==6.0.0
django-js-asset==3.1.2
django-nine==0.2.7
django-querycount==0.8.3
djangorestframework==3.16.1
drf-yasg==1.21.10
elastic-transport==9.1.0
elasticsearch-dsl==0.0.9
env==0.1.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==1.26.20
